Refactor guest auth tests and align version metadata
