+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/AiChatRuntimeTestSupport.kt
@@ -41,7 +41,7 @@ import kotlinx.coroutines.test.TestScope
 internal const val defaultTestWorkspaceId: String = "workspace-1"
 internal const val secondaryTestWorkspaceId: String = "workspace-2"
 
-private const val testAppVersion: String = "1.2.1"
+private const val testAppVersion: String = "1.2.2"
 internal const val testUiLocaleTag: String = "en-US"
 
 internal fun makeRuntime(scope: TestScope, repository: FakeAiChatRepository): AiChatRuntime {
